app/requirements.txt
pythonping==1.1.4
isodate==0.6.1
pyyaml==6.0.1
PyP100 @ git+https://github.com/almottier/TapoP100.git@main